+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
+## Unreleased
+
+- Add the experimental filled 2D circle shape builder.